# docs/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: FootballCheck
  version: '1.0'
  description: The API for the FootballCheck project
paths:
...
  "/api/v1/users":
    get:
      security:
      - Bearer: []
      tags:
      - User
      parameters:
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/UserSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserDto"
        '401':
          description: Unauthorized
    post:
      security:
      - Bearer: []
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/users/{id}":
    patch:
      security:
      - Bearer: []
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UserPasswordDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/UserDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/UserDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      security:
      - Bearer: []
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/UserDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/v1/users/register":
    post:
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserIdPasswordDto"
        '400':
          description: Bad Request
        '409':
          description: Conflict
  "/api/v1/users/login":
    post:
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/LoginDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/TokenDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/TokenDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/TokenDto"
        '400':
          description: Bad Request
        '404':
          description: Not Found
components:
  schemas:
...
    LoginDto:
      type: object
      properties:
        login:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
...
    TokenDto:
      type: object
      properties:
        accessToken:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
      additionalProperties: false
    UserDto:
      type: object
      properties:
        login:
          type: string
          nullable: true
        permission:
          type: string
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    UserIdPasswordDto:
      type: object
      properties:
        login:
          type: string
          nullable: true
        permission:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    UserPasswordDto:
      type: object
      properties:
        login:
          type: string
          nullable: true
        permission:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    UserSortState:
      enum:
      - IdAsc
      - IdDesc
      - LoginAsc
      - LoginDesc
      - PermissionAsc
      - PermissionDesc
      - RatingSquadAsc
      - RatingSquadDesc
      type: string
  securitySchemes:
    Bearer:
      type: apiKey
      description: Please insert JWT with Bearer into field
      name: Authorization
      in: header